Tsified Dot.
